requirements.txt
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-kms==2.14.0
altgraph==0.17
#bleak==0.9.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crcmod==1.7
cryptography==3.2.1
future==0.18.2
idna==2.10
importlib-metadata==3.4.0
jmespath==0.10.0
mne==0.22.0
numpy==1.19.5
oss2==2.14.0
pefile==2019.4.18
pycparser==2.20
pycryptodome==3.10.1
pyEDFlib==0.1.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pythonnet==2.5.2
pywin32-ctypes==0.2.0
requests==2.25.1
scipy==1.5.4
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
websockets==8.1
wincertstore==0.2
zipp==3.4.0
pyserial==3.5
PyMySQL==1.0.2